First pass at core challenge calculation
MVP of the basic calculation for the first part of the challenge. The biggest holdup is that the two models in the video (3d shell and 2d well) have differences which I didn't notice until I started fleshing out the math. The 3d spherical shell in the video has a non-trivial thickness, but the 2d well model indicates that we can ignore it. After a bit of research, it looks like we should use the 2d model for the first part of the challenge. For the bonus question, however, it looks like the 3d model is easier -- just use the diameter of a nucleon as the thickness of the shell.

If I feel motivated enough, I'll take a second pass at this sometime later. The Java code should be converted to proper OOP, add unit tests, etc. Not sure if it's worth the effort, but I found some code online for animated diagrams, and feel tempted to at least give it a try.